+++ sentry-ruby/spec/sentry/transport/spotlight_transport_spec.rb
@@ -6,14 +6,14 @@ RSpec.describe Sentry::SpotlightTransport do
   let(:configuration) do
     Sentry::Configuration.new.tap do |config|
       config.spotlight = true
-      config.logger = Logger.new(nil)
+      config.sdk_logger = Logger.new(nil)
     end
   end
 
   let(:custom_configuration) do
     Sentry::Configuration.new.tap do |config|
       config.spotlight = 'http://foobar@test.com'
-      config.logger = Logger.new(nil)
+      config.sdk_logger = Logger.new(nil)
     end
   end
 
@@ -27,7 +27,7 @@ RSpec.describe Sentry::SpotlightTransport do
 
   it 'logs a debug message during initialization' do
     string_io = StringIO.new
-    configuration.logger = Logger.new(string_io)
+    configuration.sdk_logger = Logger.new(string_io)
 
     subject
 
@@ -60,8 +60,8 @@ RSpec.describe Sentry::SpotlightTransport do
   describe '#send_data' do
     it 'fails a maximum of three times and logs disable once' do
       string_io = StringIO.new
-      configuration.logger = Logger.new(string_io)
-      configuration.logger.level = :debug
+      configuration.sdk_logger = Logger.new(string_io)
+      configuration.sdk_logger.level = :debug
 
       allow(::Net::HTTP).to receive(:new).and_raise(Errno::ECONNREFUSED)